Finish material saving
Move code of ToolkitUtil::Text to cc file
