+++ toolkit/assetbatcher/assetbatcher.cc
@@ -2,9 +2,6 @@
 #include "foundation/stdneb.h"
 #include "assetbatcherapp.h"
 
-#define NEBULA_TEXT_IMPLEMENT
-#include "toolkit-common/text.h"
-
 //------------------------------------------------------------------------------
 /**
 */
